requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
tavily-python
duckduckgo-search